private/make-dev-env/requirements.txt
beautifulsoup4==4.12.2
blinker==1.4
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
cliapp==1.20180812.1
cmdtest==0.32+git
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
dhash==1.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
gyp==0.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
langdetect==1.0.9
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.9.3
Markdown==3.3.6
MarkupSafe==2.1.3
more-itertools==8.10.0
mutagen==1.47.0
netifaces==0.11.0
oauthlib==3.2.0
pdf2image==1.16.3
Pillow==10.1.0
psutil==5.9.7
psycopg2==2.9.2
pybktree==1.1
pycountry==22.3.5
pycryptodomex==3.19.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pypng==0.20220715.0
python-apt==2.4.0+ubuntu2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==5.4.1
qrcode==7.4.2
redmail==0.6.0
requests==2.31.0
SecretStorage==3.3.1
selectolax==0.3.17
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
systemd-python==234
thread6==0.2.0
tornado==6.4
ttystatus==0.38
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.1.0
wadllib==1.3.6
websockets==12.0
yt-dlp==2023.11.16
zipp==1.0.0
